# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project operations
  - name: withdrawals
    description: Withdrawal operations
  - name: partners
    description: Partner (linked account) operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      description: Returns all projects with a computed final balance and withdrawal count
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project with its withdrawals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project's name and/or starting amount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project and its withdrawals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /projects/{id}/withdrawals:
    get:
      operationId: listWithdrawals
      tags: [withdrawals]
      summary: List withdrawals for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
        "404":
          description: Not found
    post:
      operationId: createWithdrawal
      tags: [withdrawals]
      summary: Add a withdrawal to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "404":
          description: Not found
  /partners/search:
    get:
      operationId: searchPartners
      tags: [partners]
      summary: Search other users who already have an account, to link as a partner
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
  /withdrawals/{id}:
    patch:
      operationId: updateWithdrawal
      tags: [withdrawals]
      summary: Update a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "404":
          description: Not found
    delete:
      operationId: deleteWithdrawal
      tags: [withdrawals]
      summary: Delete a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ProjectInput:
      type: object
      required: [name, startingAmount]
      properties:
        name:
          type: string
          minLength: 1
        startingAmount:
          type: number
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        startingAmount:
          type: number
    ProjectSummary:
      type: object
      required:
        [id, name, startingAmount, finalBalance, withdrawalCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        startingAmount:
          type: number
        finalBalance:
          type: number
        withdrawalCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProjectDetail:
      type: object
      required: [id, name, startingAmount, createdAt, withdrawals]
      properties:
        id:
          type: integer
        name:
          type: string
        startingAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
    TransactionType:
      type: string
      enum: [withdrawal, deposit]
      description: Whether the entry withdraws money from the project or adds money to it
    PaymentMethod:
      type: string
      enum:
        [
          carte_debit,
          cheque,
          virement,
          cash,
          cheque_vacances,
          carte_resto,
          autre,
        ]
      description: >
        French payment methods: debit card, cheque, bank transfer, cash,
        holiday voucher (Chèque-Vacances), meal/restaurant card (Carte
        Resto / Pass), other
    Partner:
      type: object
      required: [id, name, imageUrl]
      properties:
        id:
          type: string
          description: Clerk user id
        name:
          type: string
        email:
          type: string
        imageUrl:
          type: string
          description: Profile photo URL (e.g. from Google sign-in)
    Withdrawal:
      type: object
      required:
        [
          id,
          projectId,
          title,
          type,
          amount,
          url,
          date,
          paymentMethod,
          partner,
          createdAt,
        ]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
        url:
          type: string
        date:
          type: string
          format: date-time
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        partner:
          anyOf:
            - $ref: "#/components/schemas/Partner"
            - type: "null"
        createdAt:
          type: string
          format: date-time
    WithdrawalInput:
      type: object
      required: [title, amount, url]
      properties:
        title:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
        url:
          type: string
        date:
          type: string
          format: date-time
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        partnerUserId:
          type: string
          nullable: true
    WithdrawalUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
        url:
          type: string
        date:
          type: string
          format: date-time
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        partnerUserId:
          type: string
          nullable: true