# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Partner {
  /** Clerk user id */
  id: string;
  name: string;
  email?: string;
  /** Profile photo URL (e.g. from Google sign-in) */
  imageUrl: string;
}