# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * French payment methods: debit card, cheque, bank transfer, cash, holiday voucher (Chèque-Vacances), meal/restaurant card (Carte Resto / Pass), other
 */
export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  carte_debit: 'carte_debit',
  cheque: 'cheque',
  virement: 'virement',
  cash: 'cash',
  cheque_vacances: 'cheque_vacances',
  carte_resto: 'carte_resto',
  autre: 'autre',
} as const;