# lib/api-zod/src/generated/types/transactionType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Whether the entry withdraws money from the project or adds money to it
 */
export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  withdrawal: 'withdrawal',
  deposit: 'deposit',
} as const;