# lib/api-zod/src/generated/types/withdrawalUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethod } from './paymentMethod';
import type { TransactionType } from './transactionType';

export interface WithdrawalUpdate {
  title?: string;
  type?: TransactionType;
  amount?: number;
  url?: string;
  date?: Date;
  paymentMethod?: PaymentMethod;
  partnerUserId?: string | null;
}